+++ tests/plugins/__init__.py
@@ -3,10 +3,11 @@ from types import SimpleNamespace
 from _delb.plugins import plugin_manager
 
 
-@plugin_manager.register_document_extension
+@plugin_manager.register_document_mixin
 class TestDocumentExtension:
-    def _init_config(self, config):
-        self.config.test = SimpleNamespace(
-            initialized=True, property=config.pop("test_property", None)
+    @classmethod
+    def _init_config(cls, config, kwargs):
+        config.test = SimpleNamespace(
+            initialized=True, property=kwargs.pop("test_property", None)
         )
-        super()._init_config(config)
+        super()._init_config(config, kwargs)